ispravke

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -2,10 +2,10 @@ package ba.unsa.etf.rpr;
 
 public class Artikl {
     String naziv, kod;
-    double cijena;
+    int cijena;
 
     public Artikl() {}
-    public Artikl(String naziv, double cijena, String kod){
+    public Artikl(String naziv, int cijena, String kod){
         this.naziv=naziv;
         this.cijena=cijena;
         this.kod=kod;
@@ -31,7 +31,7 @@ public class Artikl {
         return cijena;
     }
 
-    public void setCijena(double cijena) {
+    public void setCijena(int cijena) {
         this.cijena = cijena;
     }
 }
